+++ packages/common/src/index.ts
@@ -1,4 +1,5 @@
-export * from './invitationCode'
+export * from './invitationLink/invitationLink'
+export * from './invitationLink/invitationLink.const'
 export * from './const'
 export * from './capitalize'
 export * from './process'