+++ packages/desktop/src/renderer/components/widgets/channels/ChannelMessages.test.tsx
@@ -145,7 +145,7 @@ describe('ChannelMessages', () => {
                             class="MuiGrid-root MuiGrid-item css-15myz84-MuiGrid-root"
                           >
                             <span
-                              class="MuiTypography-root MuiTypography-body1 TextMessagemessage css-3qg0or-MuiTypography-root"
+                              class="MuiTypography-root MuiTypography-body1 TextMessagemessage css-llt4u8-MuiTypography-root"
                               data-testid="messagesGroupContent-string"
                             >
                               string