+++ packages/e2e-tests/src/tests/invitationLink.test.ts
@@ -8,7 +8,7 @@ import {
   Sidebar,
   WarningModal,
 } from '../selectors'
-import { composeInvitationDeepUrl, parseInvitationCode, userJoinedMessage } from '@quiet/common'
+import { composeInvitationDeepUrl, parseInvitationLink, userJoinedMessage } from '@quiet/common'
 import { execSync } from 'child_process'
 import { type SupportedPlatformDesktop } from '@quiet/types'
 import { createLogger } from '../logger'
@@ -151,8 +151,8 @@ describe('New user joins using invitation link while having app opened', () => {
       }
 
       const copiedCode = url.hash.substring(1)
-      expect(() => parseInvitationCode(copiedCode)).not.toThrow()
-      const data = parseInvitationCode(copiedCode)
+      expect(() => parseInvitationLink(copiedCode)).not.toThrow()
+      const data = parseInvitationLink(copiedCode)
       const commandFull = `${command[process.platform as SupportedPlatformDesktop]} ${process.platform === 'win32' ? '""' : ''} "${composeInvitationDeepUrl(data)}"`
       logger.info(`Calling ${commandFull}`)
       execSync(commandFull)